requirements.txt
amply==0.1.4
appdirs==1.4.4
attrs==20.3.0
configargparse==1.3
datrie==0.8.2
docutils==0.16
gitdb==4.0.5
gitpython==3.1.13
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-core==4.7.1
nbformat==5.1.2
psutil==5.8.0
pulp==2.4
pyparsing==2.4.7
pyrsistent==0.17.3
pyyaml==5.4.1
ratelimiter==1.2.0.post0
smmap==3.0.5
snakemake==5.32.2
toposort==1.6
traitlets==5.0.5
wrapt==1.12.1
